refactor(frontend): move app-layout → features/app-shell, split header out
Bean pawrrtal-y2tt (line-budget exemption) + the same-order edges
analysis: `frontend/components/app-layout.tsx` was 569 LOC (over the
500-LOC budget) AND reached into 5 different features
(ChatActivityProvider, SidebarFocusProvider+useFocusZone, NavChats,
OnboardingModal, OnboardingFlow). It was nominally in `fe-shell`
(components/app-*) but composed `fe-features` — 6 same-order edges.

This PR moves it down to `fe-features` where the composition naturally
lives, and splits the header chrome into a sibling file:

  frontend/components/app-layout.tsx
    → frontend/features/app-shell/AppShell.tsx           (~307 LOC)
    + frontend/features/app-shell/AppShellHeader.tsx     (~280 LOC)

The export renames `AppLayout → AppShell`. `app/(app)/layout.tsx`
imports `AppShell` from `@/features/app-shell/AppShell` — this is a
clean fe-app → fe-features edge (downward, allowed). The 6 cross-feature
imports inside the file (nav-chats + onboarding) are now feature → same-
level-feature siblings, which sentrux's layer rule treats as OK.

`AppShellHeader.tsx` owns the workspace selector + history controls +
help menu + the top-bar strip — pure chrome with no nav-chats / onboarding
dependencies, so it stays small. All relative imports inside the moved
files become aliased (`@/components/...`) since the file's new home is
two directories deeper.

Doc comments referencing the old `AppLayout` name in `app-sidebar.tsx`,
`tasks/page.tsx`, `knowledge/page.tsx`, `TasksView.tsx`, and
`KnowledgeView.tsx` are updated to `AppShell`.

Verification: I can't open a browser from this env to smoke the sidebar,
onboarding modal, or chat-activity context end-to-end. The signal we have:
- file-lines budget: AppShell.tsx 307 + AppShellHeader.tsx ~280 — both well
  under 500.
- bun run typecheck: clean.
- bun run arch:check (depcruise): clean (361 modules, 1249 deps).
- vitest features/: 45 files / 212 tests, all passing.
- just sentrux: still green; quality_signal 6911 → 6909 (noise; one edge
  shifted from fe-shell → fe-features to fe-features → fe-features).

Open follow-up: file-level visual smoke (sidebar slide, onboarding open,
chat-activity context propagation) needs a `just dev` browser pass
before merging. None of the changes alter runtime behavior — they're
pure file-system relocations + name rename — so the risk is low, but a
manual click-through is the only way to be 100% sure.
