+++ frontend/components/app-sidebar.tsx
@@ -1,7 +1,7 @@
 /**
  * Thin sidebar shell that only mounts the conversations list.
  *
- * @fileoverview Prefer {@link AppLayout} for the full app chrome; this is used by the dashboard demo page.
+ * @fileoverview Prefer {@link AppShell} for the full app chrome; this is used by the dashboard demo page.
  */
 
 'use client';
